// front/src/app/interfaces/new.ts
import { SafeResourceUrl } from "@angular/platform-browser"

// export interface New {
//     title: string | ''
//     author: string | ''
//     summary: string
//     date: Date
//     url: string
// }

export interface New {
    title: string
    author: string
    date: string 
    summary: string
    url: string
}